node: organize exports into byok module
