+++ packages/node/src/actions/byok/withdraw.ts
@@ -5,7 +5,7 @@ import {
   stringifyForWasm,
 } from '@renegade-fi/core'
 import invariant from 'tiny-invariant'
-import { type Address, type BaseErrorType, toHex } from 'viem'
+import { type Address, toHex } from 'viem'
 import type { BYOKConfig } from '../../utils/createBYOKConfig.js'
 import { getBackOfQueueWallet } from './getBackOfQueueWallet.js'
 
@@ -17,8 +17,6 @@ export type WithdrawParameters = {
 }
 
 export type WithdrawReturnType = Promise<{ taskId: string }>
-
-export type WithdrawErrorType = BaseErrorType
 
 export async function withdraw(
   config: BYOKConfig,
